server/requirements.txt
flask==3.0.0
flask-socketio==5.3.5
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
flask-migrate==4.0.5
psycopg2-binary==2.9.9
aiohttp==3.9.1
beautifulsoup4==4.12.2
python-dotenv==1.0.0
gunicorn==21.2.0
